refactor: Move simulation component to separate classes
