+++ USca/USca_RTU/Processor/SimulatorNew.cs
@@ -1,254 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Drawing;
-using System.Windows.Media;
+using USca_RTU.Processor.Simulator;
 using USca_Server.Tags;
 
 namespace USca_RTU.Processor
 {
-    public partial class SimTank : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        private double _value;
-        public double Value { get { return _value; } set { _value = Math.Max(0, value); } }
-
-        public SimTank(int address, double value)
-        {
-            Address = address;
-            Value = value;
-        }
-    }
-
-    public partial class SimValve : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        public double Rate { get; set; }
-        public bool Open { get; set; }
-        public bool External { get; set; }
-
-        public SimValve(int address, double rate, bool open)
-        {
-            Address = address;
-            Rate = rate;
-            Open = open;
-            External = address <= 0;
-        }
-
-        public double Value { get { return Open ? Rate : 0; } set { Rate = value; } }
-    }
-
-    public partial class SimBindingTankValve : INotifyPropertyChanged
-    {
-        public SimTank Tank { get; set; }
-        public List<SimValve> In { get; set; } = new();
-        public List<SimValve> Out { get; set; } = new();
-
-        public SimBindingTankValve(SimTank tank, SimValve? inValve = null, SimValve? outValve = null)
-        {
-            Tank = tank;
-            if (inValve != null)
-            {
-                In.Add(inValve);
-            }
-            if (outValve != null)
-            {
-                Out.Add(outValve);
-            }
-        }
-
-        public SimBindingTankValve(SimTank tank, List<SimValve> inValve, List<SimValve> outValve)
-        {
-            Tank = tank;
-            foreach (var o in inValve) In.Add(o);
-            foreach (var o in outValve) Out.Add(o);
-        }
-
-        public void ApplyIn()
-        {
-            double diff = 0;
-            foreach (var o in In)
-            {
-                if (o.Open)
-                {
-                    diff += o.Value;
-                }
-            }
-            Tank.Value += diff;
-        }
-
-        public void ApplyOut()
-        {
-            double diff = 0;
-            foreach (var o in Out)
-            {
-                if (o.Open)
-                {
-                    diff -= o.Value;
-                }
-            }
-            Tank.Value += diff;
-        }
-    }
-
-    public partial class SimThermometer : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        public double Temperature { get; set; }
-
-        public SimThermometer(int address, double temperature)
-        {
-            Address = address;
-            Temperature = temperature;
-        }
-
-        public double Value { get { return Temperature; } set { Temperature = value; } }
-    }
-
-    public partial class SimCondenser : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        public double Rate { get; set; }
-        public bool Open { get; set; } // If not open, the 'rate' slowly goes down to 0.
-
-        public SimCondenser(int address, double rate, bool open)
-        {
-            Address = address;
-            Rate = rate;
-            Open = open;
-        }
-
-        public double Value { get { return Rate; } set { Rate = value; } }
-
-        public void Update()
-        {
-            if (!Open)
-            {
-                if (Rate > 0)
-                {
-                    Rate -= Math.Abs(Rate / 10);
-                }
-                if (Rate < 0)
-                {
-                    Rate += Math.Abs(Rate / 10);
-                }
-                if (Math.Abs(Rate) < 0.001)
-                {
-                    Rate = 0;
-                }
-            }
-        }
-    }
-
-    public partial class SimHeatSource : INotifyPropertyChanged
-    {
-        public double Rate { get; set; }
-
-        public SimHeatSource(double rate)
-        {
-            Rate = rate;
-        }
-
-        public double Value { get { return Rate; } set { Rate = value; } }
-    }
-
-    public partial class SimBindingThermometer : INotifyPropertyChanged
-    {
-        public SimThermometer Thermometer { get; set; }
-        public List<SimCondenser> Cooler { get; set; } = new();
-        public List<SimHeatSource> Heater { get; set; } = new();
-
-        public SimBindingThermometer(SimThermometer thermometer, SimCondenser? cooler = null, SimHeatSource? heater = null)
-        {
-            Thermometer = thermometer;
-            if (cooler != null)
-            {
-                Cooler.Add(cooler);
-            }
-            if (heater != null)
-            {
-                Heater.Add(heater);
-            }
-        }
-
-        public SimBindingThermometer(SimThermometer thermometer, List<SimCondenser> cooler, List<SimHeatSource> heater)
-        {
-            Thermometer = thermometer;
-            Cooler = cooler;
-            Heater = heater;
-        }
-
-        public void ApplyCooling()
-        {
-            foreach (var o in Cooler)
-            {
-                Thermometer.Temperature -= o.Value;
-            }
-        }
-
-        public void ApplyHeating()
-        {
-            foreach (var o in Heater)
-            {
-                Thermometer.Temperature += o.Value;
-            }
-        }
-    }
-
-    public partial class SimManometer : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        public double Pressure { get; set; }
-        
-        public SimManometer(int address, double pressure)
-        {
-            Address = address;
-            Pressure = pressure;
-        }
-
-        public double Value { get { return Pressure; } set { Pressure = value; } }
-    }
-
-    public partial class SimCompressor : INotifyPropertyChanged
-    {
-        public int Address { get; set; }
-        public string Name { get; set; } = "";
-        public double Rate { get; set; }
-
-        public SimCompressor(int address, double rate, bool active)
-        {
-            Address = address;
-            Rate = rate;
-        }
-
-        public double Value { get { return Rate; } set { Rate = value; } }
-    }
-
-    public partial class SimBindingPressure : INotifyPropertyChanged
-    {
-        public SimManometer Manometer { get; set; }
-        public List<SimCompressor> Compressors = new();
-
-        public SimBindingPressure(SimManometer manometer, SimCompressor compressor)
-        {
-            Manometer = manometer;
-            Compressors.Add(compressor);
-        }
-
-        public void ApplyCompression()
-        {
-            foreach (var o in Compressors)
-            {
-                Manometer.Pressure += o.Value;
-            }
-        }
-    }
-
     public partial class SimulatorNew : INotifyPropertyChanged
     {
         private int AddressCnt = 0;
